move fs-utils && AudioPlayer to common
update
