+++ platforms/runtime/platforms/cocos-play/engine/accelerometer.js
