+++ platforms/runtime/platforms/cocos-play/engine/index.js
@@ -1,8 +1,5 @@
-// polyfill
-require('./legacy-jsb-polyfill/index');
 // engine common adapter
 require('../../../common/engine/index');
 // engine platform adapter overwirte
+require('./accelerometer');
 require('./sys');
-require('./download-audio');
-require('./AudioPlayer');
