--- platforms/runtime/platforms/cocos-play/engine/legacy-jsb-polyfill/index.js
@@ -1,4 +1,0 @@
-require('./fs-utils');
-require('./canvas');
-require('./accelerometer');
-
