+++ platforms/runtime/platforms/cocos-runtime/engine/accelerometer.js
