--- platforms/runtime/platforms/cocos-runtime/engine/legacy-jsb-polyfill/canvas.js
@@ -1,6 +1,0 @@
-let originalCreateCanvas = jsb.createCanvas.bind(jsb);
-jsb.createCanvas = function () {
-    let canvas = originalCreateCanvas();
-    canvas.style = {};
-    return canvas;
-};
