+++ platforms/runtime/platforms/link-sure/engine/accelerometer.js
