+++ platforms/runtime/platforms/link-sure/engine/index.js
@@ -1,4 +1,4 @@
-require('./legacy-jsb-polyfill/index');
 require('../../../common/engine/index');
+require('./accelerometer');
 require('./sys');
 require('./resize');
