--- platforms/runtime/platforms/link-sure/engine/legacy-jsb-polyfill/index.js
@@ -1,2 +1,0 @@
-require('./accelerometer');
-require('./fs-utils');
