+++ platforms/runtime/platforms/oppo-mini-game/engine/index.js
@@ -1,6 +1,4 @@
-require('./legacy-jsb-polyfill/index');
 require('../../../common/engine/index');
+require('./accelerometer');
 require('./sys');
-require('./download-audio');
-require('./AudioPlayer');
 require('./VideoPlayer');
