--- platforms/runtime/platforms/oppo-mini-game/engine/legacy-jsb-polyfill/index.js
@@ -1,5 +1,0 @@
-require('./accelerometer')
-require('./fs-utils');
-
-// CanvasRenderingContext2D class not defined on vivo
-// window.CanvasRenderingContext2D = window.CanvasRenderingContext2D || window.__canvas.getContext('2d').constructor;
